Port Thinker's dummy weak symbol support:
Add dummy weak symbol supporting. - Since some software, like ffmpeg, involve
weak symbols, we need a way to handle it. - This patch handle weak symbol by
ignoring it at all. Since weak symbol means some symbols that is not always
existed, we can skip it when reloacte shared objects. It would be fine, since
the software should handle it when the symbols can not be found. - Without this
patch, linker would be fail to handle weak symbol when it's value is non-zero.