[bionic] abort workaround from Eclair
DO NOT MERGE

abort() does not return, which gcc interprets to mean that it doesn't
need to preserve any of the callee-save registers.  Unfortunately this
includes the link register, so if LR is used there is no way to determine
which function called abort().

We work around this by inserting a trivial stub that doesn't alter
any of the "interesting" registers and thus doesn't need to save them.
We can't just call __libc_android_abort from C because gcc uses "bl"
without first saving LR, so we use an asm statement.  This also has
the side-effect of replacing abort() with __libc_android_abort() in
the stack trace.